Re structure App for better modularity
